fix: FloatType format and id_list parsing

+++ src/main/java/com/tiger/types/FloatType.java
@@ -13,6 +13,6 @@ public class FloatType implements Type {
 
     @Override
     public String format() {
-        return "int";
+        return "float";
     }
 }
